Moved all add-on global names in the RoleBuffAddOn table.
Changed all global variables and functions so they are now placed
in the global table RoleBuffAddOn. The remaining global names
from the addon are the saved variables and the slash command
variable. Using the local name mod (or this) for the global addon
table in each of the .lua source files.